// src/common/common-module/src/enums/approvalStatus.enum.ts
export enum ApprovalStatus {
    Approved = 'Approved',
    Rejected = 'Rejected',
    Hold = 'Hold',
    ApprovedProcess = 'Approved Process',
    RejectedProcess = 'Rejected Process',
    HoldProcess = 'Hold Process',
    NotRequired = 'Not Required',
    NotRequiredProcess = 'Not Required Process'
}