// src/common/common-module/src/enums/leadStatus.enum.ts
export enum LeadStatus {
  FRESH_LEAD = 'Fresh Lead',
  CALLBACK = 'Callback',
  INTERESTED = 'Interested',
  NOT_INTERESTED = 'Not Interested',
  WRONG_NUMBER = 'Wrong Number',
  DOCUMENT_RECEIVED = 'Document Received',
  APPROVED = 'Approved',
  HOLD = 'Hold',
  DISBURSAL_SHEET_SEND = 'Disbursal Sheet Send',
  DISBURSED = 'Disbursed',
  CLOSED = 'Closed',
  PART_PAYMENT = 'Part Payment',
  SETTLEMENT = 'Settlement',
  INCOMPLETE_DOCUMENTS = 'Incomplete Documents',
  DNC = 'DNC',
  REJECTED = 'Rejected',
  NOT_ELIGIBLE = 'Not Eligible',
  DUPLICATE = 'Duplicate',
  OTHER = 'Other',
  NO_ANSWER = 'No Answer',
  EMI_PAID = 'EMI Paid',
  LESS_SALARY = 'Less Salary',
  OUT_OF_RANGE = 'Out of Range',
  EMI_PRECLOSE = 'EMI PRECLOSE',
  BANK_UPDATE_REJECTED = 'Bank Update Rejected',
  APPROVED_PROCESS = 'Approved Process',
  REJECTED_PROCESS = 'Rejected Process',
  HOLD_PROCESS = 'Hold Process',
  NOT_REQUIRED = 'Not Required',
  NOT_REQUIRED_PROCESS = 'Not Required Process',
  BLACK_LISTED = 'Blacklisted',
  DISBURSAL_APPROVED = 'Disbursal Approved',
  BANK_UPDATE_HOLD = 'Bank Update Hold',
  INCOMPLETE = 'Incomplete',
  INCOMPLETE_USER = 'Incomplete User',
}

export enum LeadType {
  NEW_CASE = 'New Case',
  REPEAT_CASE = 'Repeat Case',
  EXISTING_CASE = 'Existing Case',
}